:truck: move Ingredient entity (/src/pill > /src/ingredient)
