+++ src/main/java/com/sook/backend/pill/model/Pill.java
@@ -5,12 +5,14 @@ import java.util.List;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
+import javax.persistence.Lob;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 import com.sook.backend.attention.model.Attention;
 import com.sook.backend.attention.model.AttentionPill;
 import com.sook.backend.common.model.BaseModel;
+import com.sook.backend.ingredient.model.Ingredient;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -26,25 +28,30 @@ import lombok.NoArgsConstructor;
 @Getter
 public class Pill extends BaseModel {
 
-	private String name;
+    private String name;
 
-	@OneToMany(mappedBy = "pill", fetch = FetchType.LAZY)
-	@Builder.Default
-	private List<PillIngredient> pillIngredients = new ArrayList<>();
+    @OneToMany(mappedBy = "pill", fetch = FetchType.LAZY)
+    @Builder.Default
+    private List<PillIngredient> pillIngredients = new ArrayList<>();
 
-	@OneToMany(mappedBy = "pill", fetch = FetchType.LAZY)
-	@Builder.Default
-	private List<AttentionPill> attentionPills = new ArrayList<>();
+    @OneToMany(mappedBy = "pill", fetch = FetchType.LAZY)
+    @Builder.Default
+    private List<AttentionPill> attentionPills = new ArrayList<>();
 
-	public List<Ingredient> ingredients() {
-		return pillIngredients.stream()
-				.map(PillIngredient::ingredient)
-				.toList();
-	}
+    @Lob
+    private String image;
 
-	public List<Attention> attentions() {
-		return attentionPills.stream()
-				.map(AttentionPill::attention)
-				.toList();
-	}
+    private Integer price;
+
+    public List<Ingredient> ingredients() {
+        return pillIngredients.stream()
+                .map(PillIngredient::ingredient)
+                .toList();
+    }
+
+    public List<Attention> attentions() {
+        return attentionPills.stream()
+                .map(AttentionPill::attention)
+                .toList();
+    }
 }
